ITEMS URL IS HYPERLINKED
NOTE: The changes made to AnalysisPanel.tsx are tied to the changes made in the Orchestrator native_function.py where I separated the URL Archivo de Compra from the rest of the appended data_points in order to be able to render it in the frontend as a clickable hyperlink address. Need to change the name of each file individually.